Move internal/persistence/bucket to tributary/persistence
