+++ river/tributary/persistence_test.go
@@ -1,4 +1,4 @@
-package persistence
+package tributary
 
 import (
 	"testing"
