+++ river/tributary/tributary.go
@@ -9,8 +9,8 @@ import (
 	"golang.org/x/net/html/charset"
 	"hawx.me/code/riviera/feed"
 	"hawx.me/code/riviera/feed/common"
+	"hawx.me/code/riviera/river/data"
 	"hawx.me/code/riviera/river/events"
-	"hawx.me/code/riviera/river/internal/persistence"
 	"hawx.me/code/riviera/river/mapping"
 	"hawx.me/code/riviera/river/riverjs"
 )
@@ -43,8 +43,9 @@ type tributary struct {
 	quit    chan struct{}
 }
 
-func New(store persistence.Bucket, uri string, cacheTimeout time.Duration, mapping mapping.Mapping) *tributary {
+func New(store data.Database, uri string, cacheTimeout time.Duration, mapping mapping.Mapping) *tributary {
 	parsedUri, _ := url.Parse(uri)
+	bucket, _ := NewBucket(store, uri)
 
 	p := &tributary{
 		uri:     parsedUri,
@@ -52,7 +53,7 @@ func New(store persistence.Bucket, uri string, cacheTimeout time.Duration, mappi
 		quit:    make(chan struct{}),
 	}
 
-	p.feed = feed.New(cacheTimeout, p.itemHandler, store)
+	p.feed = feed.New(cacheTimeout, p.itemHandler, bucket)
 	p.client = &http.Client{Timeout: time.Minute, Transport: &statusTransport{http.DefaultTransport.(*http.Transport), p}}
 
 	return p
